Disable screenshot UI widgets when no dbus service
Not every system that can execute Annotator has
the hardcoded DBus service for screenshots
available. This patch disables the UI controls
for screenshots when the screenshot service is
unavailable on the system to avoid a crash.

Also disables the screenshot keyboard shortcuts.

The cli options have consciously been left to
error out in this context, as the stderr output
should be clear enough.

Signed-off-by: Antoine Mazeas <antoine@karthanis.net>